Solution to Find Eligible Age for Senior Citizen Discount

+++ Find Eligible Age for Senior Citizen Discount@@6788dd63490637cfc44e1b9d/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float age;
+    int age;
     scanf("%d", &age);
     if (age >= 60) {
         printf("Eligible");
